dma-buf: Slit the DmaBuf struct in half
Since we might not want to map the DmaBuf at all, let's split the
former DmaBuf structure into two structure, DmaBuf that is mostly a
wrapper around the Fd now, and MappedDmaBuf that provides the accessor.

A new function is provided on the DmaBuf to consume it and return a
MappedDmaBuf.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>