lib: Switch from mmap to memmap2
The mmap crate was unmaintained for 5 years, but its newer memmap and
memmap2 equivalents could only create a memory mapping from an
std::fs::File. We don't have a backing file when allocating dma-buf but
only get a file descriptor, so they were not an option.

memmap2 recently gained such support though with the 0.3 version, so
let's remove the dependency on mmap in favor of memmap2.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>