Rewrite for borrowed buffers and slices
The original code takes ownership when mapping a buffer, which doesn't fit my use case (buffers are borrowed only to prevent multiple-write problems).

This change binds read|write guards to mmaps, creating 3 kinds, one for each combination. The guards then have an as_slice() method for accessing the contents of the mmaps.

That makes it more idiomatic to read the data, compared to passing a closure.

Compared to the original, this way loses some flexibility: every change of access type between read and write requires a new mmap+munmap pair.

While this could be recovered with another layer of guards, I don't expect this to hurt my use case so I didn't implement it.

The other flexibility lost is the ability to keep a mapped buffer around. This might be worth bringing back, depending on the cost of the mmap+munmap calls.